OpsInsight-backend/requirements.txt
openai
python-dotenv
fastapi[standard]
uvicorn[standard]
requests
msal